add `BootstrapPlugin` for initialization, refactor `main.rs` to use it
