+++ src/plugins/mod.rs
@@ -1,0 +1,2 @@
+mod bootstrap;
+pub use bootstrap::*;
